Change use case page metadata, add breaks to feature name field
This page has a gigantic table that's currently broken beyond the first
few columns in Hugo. There are a large number of changes need to get it
to a readable state: this initial change breaks the page into two
sections, and adds line breaks to the feature names.

There is much additional work needed to improve this page: many of the
documentation links have the full URI and many of the same pages are
linked to repeatedly, which are signs the page could be restructured..