Rewrite "About" page and split definitions into Glossary pages
This commit rewrites the meta description (and prepends F5 as the first
instance of the name) and contents of the page. The page itself has been
made more concise, aided partially by splitting the definitions of
Ingress and Ingress controller into their own pages, which will also be
rewritten for clarity and conciseness.