Rewrite "Extensibility with NGINX Plus" page, adjust formatting
This commit rewrites the Extensibility with NGINX Plus page to be more
concise, and makes a few formatting changes - adding relative links,
updating italicized text to use double underscores, and adding a note
component regarding the usage of Annotations for Ingress resources.

It also adds a blank line at the top of each page for this feature
branch following a convention Travis is introducing to pages for his
concurrent work, which creates more breathing room for the introduction
section of each page.