Rename and edit "How NGINX Ingress Controller is Designed"
This page is an incredibly dense technical reference on how NGINX
Ingress Controller works: this commit makes a large number of changes to
make the text more concise in many places, and on the whole, improve the
product naming conventions to match the rest of the documentation set.

The commit also changes the formatting for italics from single asterisks
to underscores and in many cases removes unnecessary title attributes
from images and updates all inline links to be relative references
instead of production URLs wherever possible.