Add Ingress-NGINX Controller migration document
This commit is a close to 1:1 conversion of a blog post released on the
main NGINX website. It makes certain blocks of text somewhat more
concise and also cleans up the phrasing of certain things.

https://www.nginx.com/blog/migrating-from-community-ingress-controller-to-f5-nginx-ingress-controller/

It also adds some small improvements to a currently hidden document
comparing the featuresets of the two Ingress Controllers: it will not be
published until a later date when it can be assessed for updates, as we
do not wish to misrepresent what updates have been made to the community
Controller since that document was originally written.