Remove Migration folder, move Ingress-NGINX guide up to Installation
This commit removes the Migration folder as it's currently unnecessary:
there's only one document. If additional documents are created in the
future, then we can add the folder again.

It also changes the weight of the first set of pages and sections
correctly, to reflect the current stage of revamped documentation.